+++ flexiznam/mcms.py
@@ -1,11 +1,20 @@
 import re
-import time
+import pandas as pd
 from pymcms.main import McmsSession
 from flexiznam.config import PARAMETERS, get_password
 
 
-def get_mouse_df(mouse_name, username, password=None):
-    """Load mouse info from mcms in a dataframe"""
+def get_mouse_info(mouse_name, username, password=None):
+    """Load mouse info from mcms in a dataframe
+
+    Args:
+        mouse_name (str): Mouse name
+        username (str): Mcms username
+        password (str, optional): Mcms password. Defaults to None.
+
+    Returns:
+        dict: Mouse info
+    """
     if password is None:
         password = get_password(username=username, app="mcms")
     mcms_sess = McmsSession(username=username, password=password)
@@ -24,3 +33,37 @@ def get_mouse_df(mouse_name, username, password=None):
     if not mouse_name:
         raise IOError("Failed to download mouse info")
     return mouse_data
+
+
+def get_procedures(mouse_name, username, password=None):
+    """Load mouse procedures from mcms in a dataframe
+
+    Args:
+        mouse_name (str): Mouse name
+        username (str): Mcms username
+        password (str, optional): Mcms password. Defaults to None.
+
+    Returns:
+        dict: Mouse procedures
+    """
+    if password is None:
+        password = get_password(username=username, app="mcms")
+    mcms_sess = McmsSession(username=username, password=password)
+    procedures = mcms_sess.get_procedures(mouse_name)
+    out = []
+    for proc in procedures:
+        if proc["procedure"]["name"] == "Administration of substances into the brain under recovery anaesthesia":
+            break
+        proc_dict = {}
+        animal = proc.pop("animal")
+        assert animal["name"] == mouse_name
+        proc_dict["start_date"] = proc.pop("startDate")
+        prot = proc.pop("protocol")
+        if prot:
+            proc_dict["project_licence"] = prot.pop("projectLicence")
+            proc_dict["protocol_code"] = prot.pop("protocolCode")
+        proc = proc.pop("procedure")
+        proc_dict["procedure_name"] = proc["name"]
+        proc_dict["description"] = proc["description"]
+        out.append(proc_dict)
+    return pd.DataFrame(out)