+++ tests/test_components/test_mcms.py
@@ -7,7 +7,11 @@ USERNAME = "ab8"
 
 
 def test_get_mouse_df():
-    md = mcms.get_mouse_df(mouse_name="PZAJ2.1c", username=USERNAME)
+    md = mcms.get_mouse_info(mouse_name="PZAJ2.1c", username=USERNAME)
     assert md["mcms_id"] == 1431106
     with pytest.raises(InvalidURL):
-        mcms.get_mouse_df(mouse_name="wrongname", username=USERNAME)
+        mcms.get_mouse_info(mouse_name="wrongname", username=USERNAME)
+
+
+def test_get_procedures():
+    proc = mcms.get_procedures(mouse_name="BRAC7437.6d", username=USERNAME)